XAgentServer/requirements.txt
autogpt
beautifulsoup4
charset_normalizer
colorama
distro
docker_py
docx
duckduckgo_search
fastapi
func_timeout
GitPython
google_api_python_client
jsonlines
jsonschema
langchain
Markdown
numpy
orjson
Pillow
pinecone_client
playwright
pydantic
pylatexenc
PyPDF2
PyYAML
regex
Requests
scikit_learn
selenium
setuptools
slowapi
spacy
termcolor
tiktoken
tqdm
uvicorn
webdriver_manager
xmltodict
docstring_parser
json5
inputimeout
openai==0.28.1
yagmail
sqlalchemy
python-multipart
tenacity
pinecone-client
python-dotenv
markdown2
apscheduler
redis
pymysql
cryptography